+++ apps/TestApp/TestApp.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace TestApp;
+
+use iumioFramework\Core\Requirement\AppModel as App;
+
+/**
+ * Class TestApp
+ * @package TestApp
+ */
+
+class TestApp extends App
+{
+    /*** REGISTER TO APP ***/
+}