+++ elements/compiled/dev/smarty_compiled/6043bfbc9c46b74be4fa39e3c7a8d56cf5b9f090_0.file.index.tpl.php
@@ -1,0 +1,50 @@
+<?php
+/* Smarty version 3.1.31, created on 2017-10-22 22:06:27
+  from "/Applications/MAMP/htdocs/iumio-framework/apps/VcsvApp/Front/views/index.tpl" */
+
+/* @var Smarty_Internal_Template $_smarty_tpl */
+if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
+  'version' => '3.1.31',
+  'unifunc' => 'content_59ecfa4301a219_07437088',
+  'has_nocache_code' => false,
+  'file_dependency' => 
+  array (
+    '6043bfbc9c46b74be4fa39e3c7a8d56cf5b9f090' => 
+    array (
+      0 => '/Applications/MAMP/htdocs/iumio-framework/apps/VcsvApp/Front/views/index.tpl',
+      1 => 1508702781,
+      2 => 'file',
+    ),
+  ),
+  'includes' => 
+  array (
+  ),
+),false)) {
+function content_59ecfa4301a219_07437088 (Smarty_Internal_Template $_smarty_tpl) {
+$_smarty_tpl->_loadInheritance();
+$_smarty_tpl->inheritance->init($_smarty_tpl, true);
+?>
+
+
+<?php 
+$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_161901612759ecfa43017ac8_78735623', "parameters");
+$_smarty_tpl->inheritance->endChild($_smarty_tpl, 'template.tpl');
+}
+/* {block "parameters"} */
+class Block_161901612759ecfa43017ac8_78735623 extends Smarty_Internal_Block
+{
+public $subBlocks = array (
+  'parameters' => 
+  array (
+    0 => 'Block_161901612759ecfa43017ac8_78735623',
+  ),
+);
+public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
+?>
+
+<p>This is your app </p>
+<?php
+}
+}
+/* {/block "parameters"} */
+}