+++ elements/compiled/dev/smarty_compiled/6292899472e69d8f67c5e4695e96ad24726adad2_0.file.index.tpl.php
@@ -1,0 +1,50 @@
+<?php
+/* Smarty version 3.1.31, created on 2017-10-23 09:26:41
+  from "/Applications/MAMP/htdocs/iumio-framework/apps/TestApp/Front/views/index.tpl" */
+
+/* @var Smarty_Internal_Template $_smarty_tpl */
+if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
+  'version' => '3.1.31',
+  'unifunc' => 'content_59ed99b101ec32_03339443',
+  'has_nocache_code' => false,
+  'file_dependency' => 
+  array (
+    '6292899472e69d8f67c5e4695e96ad24726adad2' => 
+    array (
+      0 => '/Applications/MAMP/htdocs/iumio-framework/apps/TestApp/Front/views/index.tpl',
+      1 => 1508743595,
+      2 => 'file',
+    ),
+  ),
+  'includes' => 
+  array (
+  ),
+),false)) {
+function content_59ed99b101ec32_03339443 (Smarty_Internal_Template $_smarty_tpl) {
+$_smarty_tpl->_loadInheritance();
+$_smarty_tpl->inheritance->init($_smarty_tpl, true);
+?>
+
+
+<?php 
+$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_125395001959ed99b101c475_23442607', "parameters");
+$_smarty_tpl->inheritance->endChild($_smarty_tpl, 'template.tpl');
+}
+/* {block "parameters"} */
+class Block_125395001959ed99b101c475_23442607 extends Smarty_Internal_Block
+{
+public $subBlocks = array (
+  'parameters' => 
+  array (
+    0 => 'Block_125395001959ed99b101c475_23442607',
+  ),
+);
+public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
+?>
+
+<p>This is your app </p>
+<?php
+}
+}
+/* {/block "parameters"} */
+}