+++ elements/compiled/dev/smarty_compiled/9b1f951c858a680b3d2d1cea62743825d1a93440_0.file.template.tpl.php
@@ -1,0 +1,240 @@
+<?php
+/* Smarty version 3.1.31, created on 2017-10-23 09:17:56
+  from "/Applications/MAMP/htdocs/iumio-framework/apps/HelloApp/Front/views/template.tpl" */
+
+/* @var Smarty_Internal_Template $_smarty_tpl */
+if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
+  'version' => '3.1.31',
+  'unifunc' => 'content_59ed97a4df48d5_01211778',
+  'has_nocache_code' => false,
+  'file_dependency' => 
+  array (
+    '9b1f951c858a680b3d2d1cea62743825d1a93440' => 
+    array (
+      0 => '/Applications/MAMP/htdocs/iumio-framework/apps/HelloApp/Front/views/template.tpl',
+      1 => 1508743068,
+      2 => 'file',
+    ),
+  ),
+  'includes' => 
+  array (
+  ),
+),false)) {
+function content_59ed97a4df48d5_01211778 (Smarty_Internal_Template $_smarty_tpl) {
+$_smarty_tpl->_loadInheritance();
+$_smarty_tpl->inheritance->init($_smarty_tpl, false);
+?>
+<!DOCTYPE HTML>
+<!--
+	iumio Starter by TEMPLATED
+	templated.co @templatedco
+	Released for free under the Creative Commons Attribution 3.0 license (templated.co/license)
+-->
+<html>
+<head>
+    <title>iumio Starter Theme</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    
+    <link rel="stylesheet" href="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/css/index.css'),$_smarty_tpl);?>
+" />
+    <?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::fontawesomecss(array('min'=>"yes"),$_smarty_tpl);?>
+
+</head>
+
+<body>
+
+<!-- Header -->
+<header id="header">
+    <div class="inner">
+        <a href="#" class="logo">iumio Starter</a>
+        <nav id="nav">
+            <a href="#">Home</a>
+            <a href="#">Generic</a>
+            <a href="#">Elements</a>
+        </nav>
+        <a href="#navPanel" class="navPanelToggle"><span class="fa fa-bars"></span></a>
+    </div>
+</header>
+
+<!-- Banner -->
+<section id="banner" style="background-image: url('<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/banner.jpg'),$_smarty_tpl);?>
+');">
+    <h1>Welcome to iumio Starter </h1>
+    <?php 
+$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_7859629159ed97a3651c77_79906143', "parameters");
+?>
+
+
+</section>
+
+<!-- One -->
+<section id="one" class="wrapper">
+    <div class="inner">
+        <div class="flex flex-3">
+            <article>
+                <header>
+                    <h3>Magna tempus sed amet<br /> aliquam veroeros</h3>
+                </header>
+                <p>Morbi interdum mollis sapien. Sed ac risus. Phasellus lacinia, magna a ullamcorper laoreet, lectus arcu.</p>
+                <footer>
+                    <a href="#" class="button special">More</a>
+                </footer>
+            </article>
+            <article>
+                <header>
+                    <h3>Interdum lorem pulvinar<br /> adipiscing vitae</h3>
+                </header>
+                <p>Morbi interdum mollis sapien. Sed ac risus. Phasellus lacinia, magna a ullamcorper laoreet, lectus arcu.</p>
+                <footer>
+                    <a href="#" class="button special">More</a>
+                </footer>
+            </article>
+            <article>
+                <header>
+                    <h3>Libero purus magna sapien<br /> sed ullamcorper</h3>
+                </header>
+                <p>Morbi interdum mollis sapien. Sed ac risus. Phasellus lacinia, magna a ullamcorper laoreet, lectus arcu.</p>
+                <footer>
+                    <a href="#" class="button special">More</a>
+                </footer>
+            </article>
+        </div>
+    </div>
+</section>
+
+<!-- Two -->
+<section id="two" class="wrapper style1 special">
+    <div class="inner">
+        <header>
+            <h2>Ipsum Feugiat</h2>
+            <p>Semper suscipit posuere apede</p>
+        </header>
+        <div class="flex flex-4">
+            <div class="box person">
+                <div class="image round">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic03.jpg'),$_smarty_tpl);?>
+" alt="Person 1" />
+                </div>
+                <h3>Magna</h3>
+                <p>Cipdum dolor</p>
+            </div>
+            <div class="box person">
+                <div class="image round">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic04.jpg'),$_smarty_tpl);?>
+" alt="Person 2" />
+                </div>
+                <h3>Ipsum</h3>
+                <p>Vestibulum comm</p>
+            </div>
+            <div class="box person">
+                <div class="image round">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic05.jpg'),$_smarty_tpl);?>
+" alt="Person 3" />
+                </div>
+                <h3>Tempus</h3>
+                <p>Fusce pellentes</p>
+            </div>
+            <div class="box person">
+                <div class="image round">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic06.jpg'),$_smarty_tpl);?>
+" alt="Person 4" />
+                </div>
+                <h3>Dolore</h3>
+                <p>Praesent placer</p>
+            </div>
+        </div>
+    </div>
+</section>
+
+<!-- Three -->
+<section id="three" class="wrapper special">
+    <div class="inner">
+        <header class="align-center">
+            <h2>Nunc Dignissim</h2>
+            <p>Aliquam erat volutpat nam dui </p>
+        </header>
+        <div class="flex flex-2">
+            <article>
+                <div class="image fit">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic01.jpg'),$_smarty_tpl);?>
+" alt="Pic 01" />
+                </div>
+                <header>
+                    <h3>Praesent placerat magna</h3>
+                </header>
+                <p>Praesent dapibus, neque id cursus faucibus, tortor neque egestas augue, eu vulputate magna eros eu erat. Aliquam erat volutpat. Nam dui mi, tincidunt quis, accumsan porttitor lorem ipsum.</p>
+                <footer>
+                    <a href="#" class="button special">More</a>
+                </footer>
+            </article>
+            <article>
+                <div class="image fit">
+                    <img src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/images/pic02.jpg'),$_smarty_tpl);?>
+" alt="Pic 02" />
+                </div>
+                <header>
+                    <h3>Fusce pellentesque tempus</h3>
+                </header>
+                <p>Sed adipiscing ornare risus. Morbi est est, blandit sit amet, sagittis vel, euismod vel, velit. Pellentesque egestas sem. Suspendisse commodo ullamcorper magna non comodo sodales tempus.</p>
+                <footer>
+                    <a href="#" class="button special">More</a>
+                </footer>
+            </article>
+        </div>
+    </div>
+</section>
+
+<!-- Footer -->
+<footer id="footer">
+    <div class="inner">
+        <div class="flex">
+            <div class="copyright">
+                &copy; Untitled. Design: <a href="https://templated.co">TEMPLATED</a>. Images: <a href="https://unsplash.com">Unsplash</a> Modified by iumio.
+            </div>
+            <ul class="icons">
+                <li><a href="#" class="icon fa-facebook"><span class="label">Facebook</span></a></li>
+                <li><a href="#" class="icon fa-twitter"><span class="label">Twitter</span></a></li>
+                <li><a href="#" class="icon fa-linkedin"><span class="label">linkedIn</span></a></li>
+                <li><a href="#" class="icon fa-pinterest-p"><span class="label">Pinterest</span></a></li>
+                <li><a href="#" class="icon fa-vimeo"><span class="label">Vimeo</span></a></li>
+            </ul>
+        </div>
+    </div>
+</footer>
+
+<!-- Scripts -->
+<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::jquery(array(),$_smarty_tpl);?>
+
+<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::skel(array(),$_smarty_tpl);?>
+
+<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::util(array(),$_smarty_tpl);?>
+
+<?php echo '<script'; ?>
+ src="<?php echo iumioFramework\Core\Additionnal\Template\ViewBasePlugin::webassets(array('path'=>'public/js/main.js'),$_smarty_tpl);?>
+"><?php echo '</script'; ?>
+>
+
+
+
+</body>
+</html><?php }
+/* {block "parameters"} */
+class Block_7859629159ed97a3651c77_79906143 extends Smarty_Internal_Block
+{
+public $subBlocks = array (
+  'parameters' => 
+  array (
+    0 => 'Block_7859629159ed97a3651c77_79906143',
+  ),
+);
+public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
+?>
+
+        <p>Your parameter is <?php echo $_smarty_tpl->tpl_vars['sent']->value;?>
+ </p>
+    <?php
+}
+}
+/* {/block "parameters"} */
+}