+++ elements/compiled/dev/smarty_compiled/9e295fa0aa4f58caa1798e8370285e5eb1ca3320_0.file.index.tpl.php
@@ -1,0 +1,50 @@
+<?php
+/* Smarty version 3.1.31, created on 2017-10-22 22:04:57
+  from "/Applications/MAMP/htdocs/iumio-framework/apps/DzacApp/Front/views/index.tpl" */
+
+/* @var Smarty_Internal_Template $_smarty_tpl */
+if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
+  'version' => '3.1.31',
+  'unifunc' => 'content_59ecf9e9889849_82939209',
+  'has_nocache_code' => false,
+  'file_dependency' => 
+  array (
+    '9e295fa0aa4f58caa1798e8370285e5eb1ca3320' => 
+    array (
+      0 => '/Applications/MAMP/htdocs/iumio-framework/apps/DzacApp/Front/views/index.tpl',
+      1 => 1508702691,
+      2 => 'file',
+    ),
+  ),
+  'includes' => 
+  array (
+  ),
+),false)) {
+function content_59ecf9e9889849_82939209 (Smarty_Internal_Template $_smarty_tpl) {
+$_smarty_tpl->_loadInheritance();
+$_smarty_tpl->inheritance->init($_smarty_tpl, true);
+?>
+
+
+<?php 
+$_smarty_tpl->inheritance->instanceBlock($_smarty_tpl, 'Block_55568661059ecf9e9887cf9_47523588', "parameters");
+$_smarty_tpl->inheritance->endChild($_smarty_tpl, 'template.tpl');
+}
+/* {block "parameters"} */
+class Block_55568661059ecf9e9887cf9_47523588 extends Smarty_Internal_Block
+{
+public $subBlocks = array (
+  'parameters' => 
+  array (
+    0 => 'Block_55568661059ecf9e9887cf9_47523588',
+  ),
+);
+public function callBlock(Smarty_Internal_Template $_smarty_tpl) {
+?>
+
+<p>This is your app </p>
+<?php
+}
+}
+/* {/block "parameters"} */
+}